--- app/docker/components/volumesCIFSForm/volumes-cifs-form-model.js
@@ -1,10 +1,0 @@
-export function VolumesCIFSFormData() {
-  this.useCIFS = false;
-  this.serverAddress = '';
-  this.share = '';
-  this.version = 'Windows Vista / Server 2008';
-  this.versions = ['Windows XP / Server 2003 (and earlier)', 'Windows Vista / Server 2008', 'Windows 7 / Server 2008 R2', 'Windows 8 / Server 2012 (and newer)'];
-  this.versionsNumber = ['1.0', '2.0', '2.1', '3.0'];
-  this.username = '';
-  this.password = '';
-}
