+++ app/docker/components/volumesCIFSForm/volumesCifsForm.html
@@ -6,7 +6,7 @@
     </div>
     <!-- address-input -->
     <div class="form-group col-md-12">
-      <label for="cifs_address" class="col-sm-2 col-md-1 control-label text-left">Server Name/IP</label>
+      <label for="cifs_address" class="col-sm-2 col-md-1 control-label text-left">Address</label>
       <div class="col-sm-10 col-md-11">
         <input type="text" class="form-control" ng-model="$ctrl.data.serverAddress" name="cifs_address" placeholder="e.g. my.cifs-server.com OR xxx.xxx.xxx.xxx" required />
       </div>
