+++ app/docker/rest/volume.js
@@ -9,10 +9,7 @@ angular.module('portainer.docker').factory('Volume', [
     'use strict';
 
     function addVolumeNameToHeader(config) {
-      if (config.data.Name === undefined) {
-        return '';
-      }
-      return config.data.Name;
+      return config.data.Name || '';
     }
 
     return $resource(
